Move repl loop and command/script execution to nu_cli
