+++ crates/nu-cli/Cargo.toml
@@ -13,9 +13,14 @@ nu-ansi-term = "0.42.0"
 
 nu-color-config = { path = "../nu-color-config" }
 
+crossterm = "0.23.0"
+crossterm_winapi = "0.9.0"
 miette = { version = "4.1.0", features = ["fancy"] }
 thiserror = "1.0.29"
 reedline = { git = "https://github.com/nushell/reedline", branch = "main" }
 
 log = "0.4"
 is_executable = "1.0.1"
+
+[features]
+plugin = []
