+++ crates/nu-cli/src/lib.rs
@@ -1,17 +1,27 @@
+mod commands;
 mod completions;
+mod config_files;
 mod errors;
+mod eval_file;
 mod nu_highlight;
 mod print;
 mod prompt;
+mod prompt_update;
+mod reedline_config;
+mod repl;
 mod syntax_highlight;
 mod util;
 mod validation;
 
+pub use commands::evaluate_commands;
 pub use completions::NuCompleter;
 pub use errors::CliError;
+pub use eval_file::evaluate_file;
 pub use nu_highlight::NuHighlight;
 pub use print::Print;
 pub use prompt::NushellPrompt;
+pub use repl::evaluate_repl;
 pub use syntax_highlight::NuHighlighter;
 pub use util::print_pipeline_data;
+pub use util::{get_init_cwd, report_error};
 pub use validation::NuValidator;
