+++ src/main.rs
@@ -1,18 +1,12 @@
-mod commands;
-mod config_files;
-mod eval_file;
 mod logger;
-mod prompt_update;
-mod reedline_config;
-mod repl;
 mod test_bins;
 #[cfg(test)]
 mod tests;
-mod utils;
 
 use crate::logger::{configure, logger};
 use log::info;
 use miette::Result;
+use nu_cli::{evaluate_commands, evaluate_file, evaluate_repl, get_init_cwd, report_error};
 use nu_command::{create_default_context, BufferedReader};
 use nu_engine::{get_full_help, CallExt};
 use nu_parser::parse;
@@ -31,7 +25,6 @@ use std::{
         Arc,
     },
 };
-use utils::report_error;
 
 thread_local! { static IS_PERF: RefCell<bool> = RefCell::new(false) }
 
@@ -44,7 +37,7 @@ fn main() -> Result<()> {
     }));
 
     // Get initial current working directory.
-    let init_cwd = utils::get_init_cwd();
+    let init_cwd = get_init_cwd();
     let mut engine_state = create_default_context(&init_cwd);
 
     // Custom additions
@@ -187,28 +180,34 @@ fn main() -> Result<()> {
             }
 
             if let Some(commands) = &binary_args.commands {
-                #[cfg(feature = "plugin")]
-                config_files::add_plugin_file(&mut engine_state);
-
-                let ret_val = commands::evaluate(commands, &init_cwd, &mut engine_state, input, is_perf_true());
+                let ret_val = evaluate_commands(
+                    commands,
+                    &init_cwd,
+                    &mut engine_state,
+                    input,
+                    is_perf_true(),
+                );
                 if is_perf_true() {
                     info!("-c command execution {}:{}:{}", file!(), line!(), column!());
                 }
 
                 ret_val
             } else if !script_name.is_empty() && binary_args.interactive_shell.is_none() {
-                #[cfg(feature = "plugin")]
-                config_files::add_plugin_file(&mut engine_state);
-
-                let ret_val =
-                    eval_file::evaluate(script_name, &args_to_script, &mut engine_state, input, is_perf_true());
+                let ret_val = evaluate_file(
+                    script_name,
+                    &args_to_script,
+                    &mut engine_state,
+                    input,
+                    is_perf_true(),
+                );
                 if is_perf_true() {
                     info!("eval_file execution {}:{}:{}", file!(), line!(), column!());
                 }
 
                 ret_val
             } else {
-                let ret_val = repl::evaluate(&mut engine_state, binary_args.config_file, is_perf_true());
+                let ret_val =
+                    evaluate_repl(&mut engine_state, binary_args.config_file, is_perf_true());
                 if is_perf_true() {
                     info!("repl eval {}:{}:{}", file!(), line!(), column!());
                 }
